NDMF対応
