+++ Editor/AnimatorControllerMapper.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dc3415ade8c98834f9965f7afdfe71ff
+guid: a6a77f24df2e8284c9f747544ce64f28
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
