+++ Editor/NDMF/Passes.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 48db52d4a7de3094eb1aebde6d3cb35c
+guid: 6cb6891d72d22684b9255d9b33fb99cc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
